Make the tests execute properly
* Needed to define a custom bean that is used in the test, because when the annotation is not found, the default cucumber.xml was used, which made the test pass even when the annotation wasn't picked up.
* Added @IntegrationTest to glue class, so that @SpringApplicationConfiguration wasn't picked up by the meta annotation support.
