+++ spring/src/test/java/cucumber/runtime/java/spring/springapplicationconfiguration/RunCukesTest.java
@@ -5,12 +5,7 @@ import cucumber.api.junit.Cucumber;
 import org.junit.runner.RunWith;
 
 @RunWith(Cucumber.class)
-@CucumberOptions(glue = {"cucumber.runtime.java.spring.springapplicationconfiguration",
-                "cucumber.runtime.java.spring.commonglue",
-                "cucumber.api.spring"},
-        features = {"classpath:cucumber/runtime/java/spring/cukes.feature",
-                "classpath:cucumber/runtime/java/spring/xmlBasedSpring.feature",
-                "classpath:cucumber/runtime/java/spring/stepdefInjection.feature",
-                "classpath:cucumber/runtime/java/spring/transaction.feature"})
+@CucumberOptions(glue = {"cucumber.runtime.java.spring.springapplicationconfiguration"},
+        features = {"classpath:cucumber/runtime/java/spring/springBeanInjectionBoot.feature"})
 public class RunCukesTest {
 }
