--- spring/src/test/java/cucumber/runtime/java/spring/springapplicationconfiguration/WithSpringAnnotations.java
@@ -1,29 +1,0 @@
-package cucumber.runtime.java.spring.springapplicationconfiguration;
-
-import cucumber.runtime.java.spring.beans.DummyComponent;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.SpringApplicationConfiguration;
-
-@SpringApplicationConfiguration(locations = "classpath:cucumber.xml")
-public class WithSpringAnnotations {
-
-    private boolean autowired;
-
-    @Value("${cukes.test.property}")
-    private String property;
-
-    @Autowired
-    public void setAutowiredCollaborator(DummyComponent collaborator) {
-        autowired = true;
-    }
-
-    public boolean isAutowired() {
-        return autowired;
-    }
-
-    public String getProperty() {
-        return property;
-    }
-
-}
