+++ spring/src/test/resources/cucumber/runtime/java/spring/springBeanInjectionBoot.feature
@@ -1,0 +1,7 @@
+Feature: Cukes injection
+
+  Scenario: annotated bean injected
+    Then I have boot belly
+
+  Scenario: xml bean injected
+    Then I have boot belly bean
